kas_macros: move widget and make_widget impls to sub-modules
